1) VersionComponent Numeric type changed from u32 to u64; 2) unique packages with ARCH subcommand fixed

+++ src/versions_compare.rs
@@ -2,7 +2,7 @@ use std::cmp::Ordering;
 
 #[derive(Debug, Eq, PartialEq)]
 enum VersionComponent {
-    Numeric(u32),
+    Numeric(u64),
     Alphanumeric(String),
 }
